pyproject.toml
[project]
name = "ha-drone-mobile"
version = "1.0.0"
description = "Home Assistant integration for DroneMobile remote start systems"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "drone_mobile>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "pytest-homeassistant-custom-component>=0.13",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
